+++ openidconnect/CHANGELOG.md
@@ -1,3 +1,11 @@
-## 0.0.1
+## 1.0.7+1
 
-* TODO: Describe initial release.
+- Bug fix
+
+## 1.0.7
+
+- Add currentEvent Property to the client
+
+## 1.0.6
+
+- TODO: Describe initial release.